Move window related widgets to their own folder
